UI: Simplify lens system and implement ranking
Start laying the groundwork for Frames architecture

Change-type: minor
